final before mailator

+++ app/Http/Controllers/EmployeeController.php
@@ -40,7 +40,7 @@ class EmployeeController extends Controller
                             </span>
                             <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton-' . $data->id . '">
                                 <li class="dropdown-item edit" name="edit" id="' . $data->id . '">
-                                    <svg xmlns="http://www.w3.org/2000/svg" class="icon dropdown-item-icon" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                                    <svg xmlns="http://www.w3.org/2000/svg" class="icon dropdown-item-icon-edit" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
                                         <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
                                         <path d="M7 7h-1a2 2 0 0 0 -2 2v9a2 2 0 0 0 2 2h9a2 2 0 0 0 2 -2v-1" />
                                         <path d="M20.385 6.585a2.1 2.1 0 0 0 -2.97 -2.97l-8.415 8.385v3h3l8.385 -8.415z" />
@@ -49,7 +49,7 @@ class EmployeeController extends Controller
                                     Upravit zaměstnance
                                 </li>
                                 <li class="dropdown-item remove" name="remove" id="' . $data->id . '">
-                                    <svg xmlns="http://www.w3.org/2000/svg" class="icon dropdown-item-icon" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                                    <svg xmlns="http://www.w3.org/2000/svg" class="icon dropdown-item-icon-remove" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
                                         <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
                                         <path d="M15 8h.01"></path>
                                         <path d="M13 20h-6a3 3 0 0 1 -3 -3v-10a3 3 0 0 1 3 -3h10a3 3 0 0 1 3 3v6"></path>
@@ -60,7 +60,7 @@ class EmployeeController extends Controller
                                     Odstranit fotografii
                                 </li>
                                 <li class="dropdown-item delete" name="delete" id="' . $data->id . '">
-                                    <svg xmlns="http://www.w3.org/2000/svg" class="icon icon dropdown-item-icon" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                                    <svg xmlns="http://www.w3.org/2000/svg" class="icon icon dropdown-item-icon-delete" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
                                         <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
                                         <path d="M4 7h16"></path><path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12"></path>
                                         <path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3"></path><path d="M10 12l4 4m0 -4l-4 4"></path>
@@ -307,10 +307,15 @@ class EmployeeController extends Controller
     {
         $user = Employee::find($id);
         $file = $user->image;
-        $filename = public_path() . '/foto/' . $file;
-        File::delete($filename);
-        $user->delete();
-        return ['success' => true, 'message' => 'Deleted Successfully'];
+        if ($file == 'no_image.png') {
+            $user->delete();
+            return response()->json(['success' => __('User deleted successfully')]);
+        } else {
+            $filename = public_path() . '/foto/' . $file;
+            File::delete($filename);
+            $user->delete();
+            return response()->json(['success' => __('User and photo deleted successfully ')]);
+        }
     }
 
     /**
@@ -320,12 +325,14 @@ class EmployeeController extends Controller
     {
         $user = Employee::find($id);
         $file = $user->image;
-        $filename = public_path() . '/foto/' . $file;
-        File::delete($filename);
-
-        $user->image = 'no_image.png';
-        $user->save();
-
-        return ['success' => true, 'message' => 'Photo Deleted Successfully'];
+        if ($file == 'no_image.png') {
+            return response()->json(['success' => __('The default photo has been set')]);
+        } else {
+            $filename = public_path() . '/foto/' . $file;
+            File::delete($filename);
+            $user->image = 'no_image.png';
+            $user->save();
+            return response()->json(['success' => __('Photo deleted successfully')]);
+        }
     }
 }
